% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build RescueTime estimation model
Ingests a RescueTime week-level report - a data.frame with columns `very_distracting`, `distracting`,
`neutral`, `productive` and `very_productive` - and returns a fitted `stanfit` model object conditioned
on the week-level productivity-level proportions contained within.}
\usage{
buildModel(report, chains = 1, cores = 1, iter = 1000, warmup = 1000)
}
\arguments{
\item{report}{A data-frame of week-level RescueTime productivity-level proportions}

\item{chains}{The number of chains to use in the Stan sampler}

\item{cores}{The number of cores to use in the Stan sampler}

\item{iter}{The number of iterations to use in the Stan sampler}

\item{warmup}{The number of warmup iterations to use in the Stan sampler}
}
\value{
The fitted `stanfit` model object
}
\description{
Build RescueTime estimation model
Ingests a RescueTime week-level report - a data.frame with columns `very_distracting`, `distracting`,
`neutral`, `productive` and `very_productive` - and returns a fitted `stanfit` model object conditioned
on the week-level productivity-level proportions contained within.
}
