% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RescueTimeEstimator-package.R
\docType{package}
\name{RescueTimeEstimator-package}
\alias{RescueTimeEstimator-package}
\alias{RescueTimeEstimator}
\title{RescueTime Estimation via the "Poor Man's Dirichlet"}
\description{
RescueTime Estimation via the "Poor Man's Dirichlet"
}
